+++ src/main/resources/db/migration/V0__init_tables.sql
@@ -10,8 +10,10 @@ create table if not exists cloud.files
     id                 SERIAL primary key,
     name               varchar(255) not null,
     creation_date_time TIMESTAMPTZ DEFAULT Now(),
-    user_id            bigint       not null,
-    size_in_bytes      int          not null,
+--     user_id            bigint       not null,
+    user_id            int       not null,
+--     size_in_bytes      int          not null,
+    size_in_bytes      bigint          not null,
     content            oid          not null,
     content_type       varchar(255) not null,
     FOREIGN KEY (user_id) references cloud.users (id)
